Media Editor Modal: Try placing the save and cancel buttons in the footer (#78708)
* Media Editor Modal: Try out placing the save and cancel buttons in the footer

* Update tsconfig

* Reorganise things a little better

* Try breaking the controls up into separate groups to allow three different breakpoint arrangements

* Add wrap if it's really needed
