+++ packages/media-editor/src/components/media-editor/style.scss
@@ -33,15 +33,6 @@
 		flex: 1;
 	}
 
-	.interface-interface-skeleton__footer {
-		position: relative;
-		display: flex;
-	}
-
-	.interface-interface-skeleton.has-footer .interface-interface-skeleton__body {
-		padding-bottom: 0;
-	}
-
 	.media-editor__canvas {
 		display: flex;
 		align-items: center;
