+++ packages/media-editor/tsconfig.json
@@ -8,6 +8,7 @@
 	"references": [
 		{ "path": "../api-fetch" },
 		{ "path": "../components" },
+		{ "path": "../compose" },
 		{ "path": "../core-data" },
 		{ "path": "../data" },
 		{ "path": "../dataviews" },
